The following approaches were used to speed up the process.

* It used to send every 1 byte, but it now holds a maximum of 63 bytes before sending
*  if the buffer is full, send buffer
* If there is data at the time of the handleUSBIRQ every 1ms, send buffer

Unlike #1452, it does not require additional memory.
The code is a little more complicated, but it looks like it works very well.

* 13 byte : fmt.Printf("hello world\r\n")
  * old : 202.814 us
  * new :  51.330 us
* 11 byte : println("hello world")
  * old : 171.274 us
  * new :  11.920 us
* 1024 byte + "\r\n"
  * old : 13.999962 ms
  * new :  1.150692 ms
* 1024 * 10 byte + "\r\n"
  * old : 139.649294 ms
  * new :  11.612682 ms

`interrupt.Disable()` cannot be built unless interp-rewrite is merged.

Samd2x can be made faster in the same way.
Once we have reviewed this PR, we will create a PR.