@sago35 2020-11-04
The reason why it is set to draft is that we need to merge the interp-rewrite branches to be able to build.
I would like a review.

I tested it with the following code with WioTerminal.

```go
package main

import (
	"fmt"
	"machine"
	"time"
)

var buf = [1024 * 10]byte{}

var (
	debugPort1 = machine.BCM3
)

func main() {
	debugPort1.Configure(machine.PinConfig{Mode: machine.PinOutput})

	for i := range buf {
		if ((i + 1) & 0x0F) == 0 {
			buf[i] = byte(fmt.Sprintf("%X", byte(i>>4)&0xF)[0])
		} else {
			buf[i] = '.'
		}
	}
	buf[0] = '+'

	for {
		debugPort1.High()

		machine.UART0.Write(buf[:1024*10])
		//machine.UART0.Write(buf[:1024])
		machine.UART0.Write([]byte("\r\n"))
		//fmt.Printf("hello world\r\n")
		//println("hello world")

		debugPort1.Low()
		time.Sleep(time.Second)
	}
}

```