@sago35 2020-11-04
* Use udd_ep_in_cache_buffer[usb_CDC_ENDPOINT_IN][128] as two 64 byte buffers.
  * The term udd_ep_in_cache_buffer is abbreviated to buffer below
* WriteByte()
  * Set `c byte` to buffer[TxIdx].
  * Can write up to 63 bytes.
  * The reason it's 63 bytes instead of 64 bytes is because `TxIdx & 0x3F` treats it as a size
  * Call Flush() if you've written to the maximum
* Flush()
  * Flush() sends the data in the buffer
  * Calculate the size to be sent using `TxIdx & 0x3F`
  * When the size is greater than 0, it is actually sent to USBCDC
  * After sending, set TxIdx to the next buffer
    * For example, when TxIdx is 0 to 63, TxIdx is set to 64. And Set TxIdx to 0 when the value is 64 to 127.
  * Set UART0.sent to true if it is sent once.
    * If not using UART0.sent, USBCDC status is not known at WriteByte() time.
* handleUSBIRQ()
  * Interrupts every 1ms
  * Calling Flush()