@sago35 2021-02-14
It is buildable on the current dev branch so it is ready for review.

I wish I could make the implementation more straightforward, but I can't think of a good way to do it right now.

The code for testing can be found at
https://github.com/tinygo-org/tinygo/pull/1481#issuecomment-721702039

I checked it on Windows 10 and found the following

* 1024 * 10 byte + "\r\n"
  * old :  514.359 ms
  * new : 11.633ms