@deadprogram 2021-02-14
The code does appear to work. I made a few comments about using `const` values for clarity.

@aykevl do you have any feedback?