@deadprogram 2021-02-15
Made a few small comments @sago35 that look like we can shorten by a couple of lines of code.