@sago35 2021-02-15
Updated. @deadprogram 

I thought USBCDC.Flush() should not be changed.
Should I change it?