@deadprogram 2021-02-15
This is really great, thank you for working on it @sago35 just waiting for CI to finish to merge.

Now we just need the same thing for SAMD21! :smile_cat: